requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-project==0.8.3
appdirs==1.4.4
argh==0.26.2
argon2-cffi==20.1.0
asn1crypto==1.4.0
astroid==2.4.2
astropy==4.0.2
async-generator==1.10
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.3.0
audeer==1.16.0
audiofile==0.4.2
audioread==2.1.9
audtorch==0.6.4
autopep8==1.5.4
Babel==2.8.1
backcall==0.2.0
backports.shutil-get-terminal-size==1.0.0
basic-model==0.4.2
beautifulsoup4==4.9.3
bidict==0.22.1
biosppy==1.0.0
bitarray==1.6.1
bitmath==1.3.3.1
bkcharts==0.2
bleach==3.2.1
bokeh==2.2.3
boto==2.49.0
Bottleneck==1.3.2
brotlipy==0.7.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.3
chardet==3.0.4
charset-normalizer==2.0.3
click==7.1.2
cloudpickle==1.6.0
clyent==1.2.2
cmake==3.22.5
colorama==0.4.4
comet-ml==3.33.11
configobj==5.0.8
configparser==5.0.2
contextlib2==0.6.0.post1
cryptography==3.1.1
cycler==0.10.0
Cython==0.29.21
cytoolz==0.11.0
dask==2.30.0
dataclasses==0.6
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
distributed==2.30.1
docker-pycreds==0.4.0
docutils==0.16
dulwich==0.21.6
entrypoints==0.3
et-xmlfile==1.0.1
everett==3.1.0
facenet-pytorch==2.5.2
fastcache==1.1.0
filelock==3.0.12
flake8==3.8.4
Flask==1.1.2
fsspec==2022.3.0
future==0.18.2
gevent==20.9.0
gitdb==4.0.7
GitPython==3.1.19
glob2==0.7
gmpy2==2.0.8
google-auth==1.32.1
google-auth-oauthlib==0.4.4
graphql-core==3.1.5
greenlet==0.4.17
grpcio==1.38.1
h5py==2.10.0
HeapDict==1.0.1
html5lib==1.1
humanize==4.0.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.8
imagesize==1.2.0
importlib-metadata==2.0.0
iniconfig==1.1.1
intervaltree==3.1.0
ipykernel==5.3.4
ipython==7.19.0
ipython_genutils==0.2.0
ipywidgets==7.5.1
iso-639==0.4.5
isort==5.6.4
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.17.1
jeepney==0.5.0
Jinja2==2.11.2
joblib==0.17.0
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab==2.2.6
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
keyring==21.4.0
kiwisolver==1.3.0
lazy-object-proxy==1.4.3
libarchive-c==2.9
librosa==0.8.0
lightning-bolts==0.3.4
llvmlite==0.34.0
lmdb==1.3.0
locket==0.2.0
lxml==4.6.1
Markdown==3.3.4
markdown-it-py==2.2.0
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
mdurl==0.1.2
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
mock==4.0.2
more-itertools==8.6.0
moviepy==1.0.3
mpmath==1.1.0
msgpack==1.0.0
multidict==5.1.0
multipledispatch==0.6.0
natsort==8.0.2
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.2
networkx==2.5
nltk==3.5
nose==1.3.7
notebook==6.1.4
numba==0.51.2
numexpr==2.7.1
numpy==1.21.5
numpydoc==1.1.0
oauthlib==3.1.1
olefile==0.46
opencv-python==4.4.0.46
openpyxl==3.0.5
packaging==20.4
pandas==1.1.3
pandocfilters==1.4.3
parso==0.7.0
partd==1.1.0
path==15.0.0
pathlib2==2.3.5
pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pip==21.2.2
pkginfo==1.6.1
pluggy==0.13.1
ply==3.11
pooch==1.4.0
proglog==0.1.10
progress==1.6
prometheus-client==0.8.0
promise==2.3
prompt-toolkit==3.0.8
protobuf==3.17.3
psutil==5.7.2
ptyprocess==0.6.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycosat==0.6.3
pycparser==2.20
pycrypto==2.6.1
pycurl==7.43.0.6
pyDeprecate==0.3.0
pydocstyle==5.1.1
pydub==0.25.1
pyflakes==2.2.0
Pygments==2.16.1
pylint==2.6.0
pymkv==1.0.8
pymongo==4.1.1
pyodbc==4.0.0-unsupported
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
pytest==0.0.0
python-box==6.1.0
python-dateutil==2.8.0
python-jsonrpc-server==0.4.0
python-language-server==0.35.1
python-speech-features==0.6
pytorch-lightning==1.3.8
pytz==2020.1
PyWavelets==1.1.1
pyxdg==0.27
PyYAML==5.3.1
pyzmq==19.0.2
QDarkStyle==2.8.1
QtAwesome==1.0.1
qtconsole==4.7.7
QtPy==1.9.0
regex==2020.10.15
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==1.0.0
resampy==0.2.2
responses==0.21.0
rich==13.6.0
rope==0.18.0
rsa==4.7.2
Rtree==0.9.4
ruamel_yaml==0.15.87
schema==0.7.5
scikit-image==0.17.2
scikit-learn==1.0.2
scipy==1.5.2
seaborn==0.11.0
SecretStorage==3.1.2
semantic-version==2.10.0
Send2Trash==1.5.0
sentry-sdk==1.3.0
setuptools==50.3.1.post20201107
shortuuid==1.0.1
simplegeneric==0.8.1
simplejson==3.17.6
singledispatch==3.4.0.3
six==1.12.0
smmap==4.0.0
snowballstemmer==2.0.0
sortedcollections==1.2.1
sortedcontainers==2.2.2
SoundFile==0.10.3.post1
soupsieve==2.0.1
sox==1.4.1
Sphinx==3.2.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-websupport==1.2.4
spyder==4.1.5
spyder-kernels==1.9.4
SQLAlchemy==1.3.20
statsmodels==0.12.0
subprocess32==3.5.4
sympy==1.6.2
tables==3.6.1
tabulate==0.8.9
tblib==1.7.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.10.1
toml==0.10.1
toolz==0.11.1
torch==1.9.0
torchaudio==0.9.0a0+33b2469
torchmetrics==0.4.1
torchsummary==1.5.1
torchvision==0.10.0
tornado==6.0.4
tqdm==4.50.2
traitlets==5.0.5
typed-ast==1.4.1
typeguard==2.13.3
typing==3.6.4
typing_extensions==4.2.0
ujson==4.0.1
unicodecsv==0.14.1
urllib3==1.26.6
wandb==0.12.10
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
websockets==10.3
Werkzeug==1.0.1
wheel==0.35.1
widgetsnbextension==3.5.1
wrapt==1.11.2
wurlitzer==2.0.1
xlrd==1.2.0
XlsxWriter==1.3.7
xlwt==1.3.0
yapf==0.30.0
yarl==1.6.3
yaspin==2.1.0
zict==2.0.0
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.1.2